Restructure stays folder for organization
