+++ backend/src/models/stays/index.ts
@@ -1,0 +1,15 @@
+import {
+  findStaySummary,
+  findAllStays,
+  postStayData,
+  putStayData,
+  deleteStayData,
+} from './stays';
+
+export {
+  deleteStayData,
+  findAllStays,
+  findStaySummary,
+  postStayData,
+  putStayData,
+};
